I made changes to the backend and the front end.
For consistency's sake, we should use the terms "participant" and "operator".

When someone loads / or /operator, by default the send button is disabled.
After connecting to the server with socket.io, each interface identifies itself as either "participant" or "operator".
The server will then respond with a status update according to which roles were already filled.

As soon as both the operator and participant are present, the participant's send button is enabled so they can ask a question.
Once their question is asked, the send button is disabled again and the operator's send button is enabled.
Once the operator has submitted their reply, the participant is given both responses and their send button is enabled again.

Meanwhile, the description at the top of the page is being modified by the server at each step to reflect the current state
of the turing test activity.